refactor(rust-sys)!: remove the deprecated async design
